refactor(auth): :lipstick: change route name with french words
