+++ src/app/(main-site)/(auth)/creation-compte/page.tsx
@@ -2,7 +2,7 @@
 
 import { useTranslation } from '@/i18n/useTranslation';
 import { useUser } from '@/lib/contexts/UserContext';
-import { Email, Lock, Visibility, VisibilityOff } from '@mui/icons-material';
+import { Email, Lock, Person, Visibility, VisibilityOff } from '@mui/icons-material';
 import {
   Alert,
   Box,
@@ -19,26 +19,38 @@ import {
   Typography,
   useMediaQuery
 } from '@mui/material';
-import { Field, Form, Formik } from 'formik';
+import { Form, Formik } from 'formik';
 import Image from 'next/image';
 import Link from 'next/link';
 import { useState } from 'react';
 import * as Yup from 'yup';
 
 // Schéma de validation
-const LoginSchema = Yup.object().shape({
+const RegisterSchema = Yup.object().shape({
+  firstName: Yup.string()
+    .required('Prénom requis'),
+  lastName: Yup.string()
+    .required('Nom requis'),
   email: Yup.string()
     .email('Email invalide')
     .required('Email requis'),
   password: Yup.string()
-    .min(6, 'Le mot de passe doit contenir au moins 6 caractères')
+    .min(8, 'Le mot de passe doit contenir au moins 8 caractères')
+    .matches(
+      /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/,
+      'Le mot de passe doit contenir au moins une minuscule, une majuscule, un chiffre et un caractère spécial'
+    )
     .required('Mot de passe requis'),
+  confirmPassword: Yup.string()
+    .oneOf([Yup.ref('password')], 'Les mots de passe doivent correspondre')
+    .required('Confirmation du mot de passe requise'),
 });
 
-export default function LoginPage() {
-  const { login, isLoading, error } = useUser();
+export default function RegisterPage() {
+  const { register, isLoading, error } = useUser();
   const { t } = useTranslation();
   const [showPassword, setShowPassword] = useState(false);
+  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
 
   // Détection du mode sombre du système
   const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
@@ -72,6 +84,10 @@ export default function LoginPage() {
     setShowPassword(!showPassword);
   };
 
+  const handleToggleConfirmPasswordVisibility = () => {
+    setShowConfirmPassword(!showConfirmPassword);
+  };
+
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
@@ -102,7 +118,7 @@ export default function LoginPage() {
         </Box>
 
         {/* Contenu du formulaire */}
-        <Container maxWidth="sm" sx={{
+        <Container maxWidth="md" sx={{
           py: 8,
           minHeight: '100vh',
           display: 'flex',
@@ -124,10 +140,10 @@ export default function LoginPage() {
           >
             <Box sx={{ textAlign: 'center', mb: 4 }}>
               <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 'bold' }}>
-                {t('auth.login')}
+                {t('auth.register')}
               </Typography>
               <Typography variant="body1" color="textSecondary">
-                {t('auth.login_instructions')}
+                {t('auth.register_instructions')}
               </Typography>
             </Box>
 
@@ -138,14 +154,79 @@ export default function LoginPage() {
             )}
 
             <Formik
-              initialValues={{ email: '', password: '' }}
-              validationSchema={LoginSchema}
+              initialValues={{
+                firstName: '',
+                lastName: '',
+                email: '',
+                password: '',
+                confirmPassword: ''
+              }}
+              validationSchema={RegisterSchema}
               onSubmit={async (values) => {
-                await login(values.email, values.password);
+                await register(
+                  values.firstName,
+                  values.lastName,
+                  values.email,
+                  values.password
+                );
               }}
             >
               {({ errors, touched, handleChange, handleBlur, values }) => (
                 <Form>
+                  <Box
+                    sx={{
+                      display: 'flex',
+                      flexDirection: { xs: 'column', sm: 'row' },
+                      gap: 2,
+                      mb: 3
+                    }}
+                  >
+                    <Box sx={{ flex: 1 }}>
+                      <TextField
+                        fullWidth
+                        id="firstName"
+                        name="firstName"
+                        type="text"
+                        label={t('auth.first_name')}
+                        variant="outlined"
+                        value={values.firstName}
+                        onChange={handleChange}
+                        onBlur={handleBlur}
+                        error={Boolean(errors.firstName && touched.firstName)}
+                        helperText={(errors.firstName && touched.firstName) ? errors.firstName : ''}
+                        InputProps={{
+                          startAdornment: (
+                            <InputAdornment position="start">
+                              <Person color="action" />
+                            </InputAdornment>
+                          ),
+                        }}
+                      />
+                    </Box>
+                    <Box sx={{ flex: 1 }}>
+                      <TextField
+                        fullWidth
+                        id="lastName"
+                        name="lastName"
+                        type="text"
+                        label={t('auth.last_name')}
+                        variant="outlined"
+                        value={values.lastName}
+                        onChange={handleChange}
+                        onBlur={handleBlur}
+                        error={Boolean(errors.lastName && touched.lastName)}
+                        helperText={(errors.lastName && touched.lastName) ? errors.lastName : ''}
+                        InputProps={{
+                          startAdornment: (
+                            <InputAdornment position="start">
+                              <Person color="action" />
+                            </InputAdornment>
+                          ),
+                        }}
+                      />
+                    </Box>
+                  </Box>
+
                   <Box mb={3}>
                     <TextField
                       fullWidth
@@ -203,33 +284,38 @@ export default function LoginPage() {
                     />
                   </Box>
 
-                  <Box mb={3} display="flex" justifyContent="space-between" alignItems="center">
-                    <Box display="flex" alignItems="center">
-                      <Field
-                        type="checkbox"
-                        name="remember"
-                        id="remember"
-                        className="h-4 w-4 mr-2"
-                      />
-                      <Typography variant="body2" component="label" htmlFor="remember">
-                        {t('auth.remember_me')}
-                      </Typography>
-                    </Box>
-                    <Link href="/forgot-password" passHref>
-                      <Typography
-                        variant="body2"
-                        component="span"
-                        color="primary"
-                        sx={{
-                          cursor: 'pointer',
-                          '&:hover': {
-                            textDecoration: 'underline'
-                          }
-                        }}
-                      >
-                        {t('auth.forgot_password')}
-                      </Typography>
-                    </Link>
+                  <Box mb={4}>
+                    <TextField
+                      fullWidth
+                      id="confirmPassword"
+                      name="confirmPassword"
+                      type={showConfirmPassword ? "text" : "password"}
+                      label={t('auth.confirm_password')}
+                      variant="outlined"
+                      value={values.confirmPassword}
+                      onChange={handleChange}
+                      onBlur={handleBlur}
+                      error={Boolean(errors.confirmPassword && touched.confirmPassword)}
+                      helperText={(errors.confirmPassword && touched.confirmPassword) ? errors.confirmPassword : ''}
+                      InputProps={{
+                        startAdornment: (
+                          <InputAdornment position="start">
+                            <Lock color="action" />
+                          </InputAdornment>
+                        ),
+                        endAdornment: (
+                          <InputAdornment position="end">
+                            <IconButton
+                              onClick={handleToggleConfirmPasswordVisibility}
+                              edge="end"
+                              aria-label="toggle password visibility"
+                            >
+                              {showConfirmPassword ? <VisibilityOff /> : <Visibility />}
+                            </IconButton>
+                          </InputAdornment>
+                        ),
+                      }}
+                    />
                   </Box>
 
                   <Button
@@ -247,7 +333,7 @@ export default function LoginPage() {
                     {isLoading ? (
                       <CircularProgress size={24} color="inherit" />
                     ) : (
-                      t('auth.login_button')
+                      t('auth.register_button')
                     )}
                   </Button>
                 </Form>
@@ -256,8 +342,8 @@ export default function LoginPage() {
 
             <Box mt={4} textAlign="center">
               <Typography variant="body2" color="textSecondary">
-                {t('auth.no_account')}{' '}
-                <Link href="/register" passHref>
+                {t('auth.already_have_account')}{' '}
+                <Link href="/connexion" passHref>
                   <Typography
                     variant="body2"
                     component="span"
@@ -270,7 +356,7 @@ export default function LoginPage() {
                       }
                     }}
                   >
-                    {t('auth.create_account')}
+                    {t('auth.login_here')}
                   </Typography>
                 </Link>
               </Typography>
